backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfminer.six==20221105
python-docx==1.1.2
spacy==3.8.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
sentence-transformers==3.0.0
networkx==3.3
numpy==1.26.4
pydantic==2.7.1
rapidfuzz>=3.6.0
